+++ app/src/main/java/com/copixelate/ui/util/InputValidation.kt
@@ -1,4 +1,4 @@
-package com.copixelate.data
+package com.copixelate.ui.util
 
 import android.util.Patterns
