+++ app/src/main/java/com/copixelate/viewmodel/ArtViewModel.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import android.util.Log
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.copixelate.data.ArtRepo
+import com.copixelate.data.repo.ArtRepo
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
